Error in getTEvents
It seems this error goes all the way back to the code provided in the book.  Really appreciate you making your examples available.
   
sNeg is off by a minus sign causing it to be reset to 0 every loop so sNeg<-h is never executed.  Events are only being triggered on the positive side (sPos).  Subtracting diff.loc[i] seems to be the easiest fix but you might prefer a different way.